refactor(cmd/riffpad): shrink main.go to dispatch-only (129 lines)
Issue #282 step 6 / acceptance criterion. main.go now only:

- parses the global --lang flag and propagates the i18n bundle
  (commands/console SetBundle)
- resolves RIFFPAD_URL/RIFFPAD_DIR and injects the data dir into cliutil
- installs the config.OnHeal hook
- dispatches to commands.* (incl. RunDaemon for the hidden _daemon entry,
  moved with enrichDaemonPath to internal/commands/rundaemon.go)
- prints usage and exits with the conventional codes

All one-line wrappers are gone; the dispatch switch calls the commands
package directly. Integration smoke test moves to internal/commands.

End-to-end verified against a real daemon: daemon start → status →
pair --local → sessions → run --cli demo → kill → daemon stop.
